queue: refactor `PusherBlocking` and `PusherDropping` to `Push` and `TryPush`
